requirements.txt
tensorflow-gpu
sklearn
pillow
pandas
numpy
torch
